third_party/WebKit/LayoutTests/shadow-dom/v0/closed-mode-deep-combinators-expected.txt

PASS result.length is 7
PASS node.id is "openhost"
PASS node.id is "closedhost"
PASS node.id is "v0host"
PASS node.id is "v0host_in_v0host"
PASS node.id is "div_v0_in_v0"
PASS node.id is "openhost_in_v0host"
PASS node.id is "closedhost_in_v0host"
PASS successfullyParsed is true

TEST COMPLETE
